Fix folding for lines that contain a comment
In attempting to optimize the performance of `isFoldableAtBufferRow` in
cd54382, we mistakenly introduced a bug that caused lines that contained
a comment but didn't start with one to not be foldable anymore.

With this commit we are restoring the previous behavior, thus only
disabling folding for lines that start with a comment (ignoring leading
whitespaces).