Set the tabIndex on the input element
This commit reverts a portion of the changes introduced in
b64cd48dd43012fbf4573e376108fe53aa6a5576. Prior to that commit, we were
setting the tabIndex to -1 on the atom-text-editor element. This commit
restores that behavior.

Instead of setting a custom tab index directly on the atom-text-editor
element, we instead set the tabIndex on the input element *inside* the
atom-text-editor element.

With these changes in place, you can successfully use the tabIndex to
define the tab order for atom-text-editor elements. 😅