Update StyleManager specs to address deprecated ::shadow selectors
As of Chromium 63, ::shadow CSS selectors have been fully removed:

https://developers.google.com/web/updates/2017/10/remove-shadow-piercing

As a result, any usage of ::shadow in a stylesheet seems to get removed as soon
as the stylesheet is applied to an HTMLStyleElement.  Since we have a couple of tests
that expect ::shadow selectors to be present, we must remove those expectations
since those selectors do not appear any longer.