Remove logic supporting dynamic switching of path watcher backends
The logic is pretty complex and I don't want to take the time to 
integrate @atom/notify with it. I left a bunch of stuff commented out in 
this commit just in case these changes break the build. I'll do another 
pass to delete commented code once we go green.