Allow escaped " literals in the SyntaxScopeMap
In order to pass a quote literal to the postcss-selector-parser, it 
needs to be escaped. However, this escaping is not removed by the parser 
in the yielded value token, so in this commit I replace any escaped 
quote literals with unescaped quotes after we trim the outer quotes off 
of the string.