Don't symlink compile-cache folder in AtomWindow tests
In these tests, we create a temporary `ATOM_HOME` to avoid cluttering 
the user's real `~/.atom` folder.

Adding a symlink to the real `compile-cache` was introduced to speed up 
main process tests, so that the transpilation cache could be reused. 

Unfortunately, when the real `~/.atom` folder did not exist (such as on 
a pristine environment on CI), it would confuse Atom, which would think 
that it didn't need to re-create a `compile-cache` folder again, but 
wouldn't be able to write to it because the symlink pointed to a 
non-existant directory.

Main process tests were overhauled and made faster recently, so we can 
safely remove this performance optimization.