Do not depend on CONFIG.executableName on the windows tests
CONFIG.executableName depends on the ATOM_RELEASE_VERSION env variable
to work correctly on nightly releases.

Since this env variable is not being set for the testing steps, the
nightly release builds contain a wrong executableName on the testing
step, which causes a failure when trying to find the Atom executable.

This commit fixes that by stop using the executableName from the test
step, like it's done on other platforms.